# specs/008-duckdock-2-foundation/contracts/openapi-v2.yaml
openapi: 3.1.0
...
info:
  title: DuckDock Agent Execution API
  version: 2.0.0-draft
  summary: Reporter ingestion and governed AgentRun read contracts for DuckDock 2.0.
  description: |
    Foundation contract for metadata-only Agent Harness execution reporting.
    Reporter writes are authenticated with scoped Reporter bearer tokens and are
    idempotent. Namespace and Runtime identity are derived from the credential;
    they are never accepted from the Reporter request body. Raw prompts,
    completions, tool arguments, tool results, retrieved content, and hidden
    chain-of-thought are outside this API contract.
servers:
  - url: /
...
paths:
...
  /api/v2/reporter/telemetry-sinks/{sink_public_id}/v1/traces:
    post:
      tags:
        - Generic OTLP
      summary: Reconcile sanitized Generic OTLP traces
      description: |
        Accepts uncompressed standard OTLP/HTTP JSON from the registered
        per-Runtime Collector. Namespace, Runtime and trust are derived from
        the execution.write ReporterCredential. DuckDock persists only
        metadata projections; raw spans and attributes remain in the configured
        trace backend. Ambiguous or conflicting mappings return OTLP partial
        success and enter quarantine.
      operationId: ingestGenericOtlpTraces
      security:
        - ReporterBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TelemetrySinkProjectionPublicId'
        - $ref: '#/components/parameters/AdapterHandshakeHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtlpTraceExportRequest'
      responses:
        '200':
          description: Export accepted, optionally with per-batch partial success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtlpTraceExportResponse'
        '400':
          description: Malformed or unsupported OTLP JSON.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          description: OTLP projection request exceeds 1 MiB.
        '415':
          description: Only uncompressed application/json is accepted.
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /api/v2/reporter/handshakes:
    post:
      tags:
        - Adapter Fleet
      summary: Negotiate one credential-bound adapter descriptor
      operationId: negotiateAdapterHandshake
      security:
        - ReporterBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdapterDescriptor'
      responses:
        '201':
          description: Dynamic capabilities accepted and rejected by current dependencies.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdapterHandshake'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v2/reporter/heartbeats:
    post:
      tags:
        - Adapter Fleet
      summary: Record adapter health, Collector version and configuration drift
      operationId: recordAdapterHeartbeat
      security:
        - ReporterBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdapterHeartbeat'
      responses:
        '200':
          description: Heartbeat accepted and compared with the negotiated descriptor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdapterHeartbeatAck'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v2/fleet/runtimes:
    get:
      tags:
        - Adapter Fleet
      summary: Read dynamic Runtime capability and heartbeat Fleet state
      operationId: getFleetRuntimeSummary
      security:
        - UserBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/NamespaceId'
      responses:
        '200':
          description: Runtime Fleet state visible to one Namespace member.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FleetSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    ReporterBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: DuckDock Reporter Token
      description: |
        Scoped, rotatable Reporter credential bound to a Namespace and Runtime.
        The execution API requires the `execution.write` scope. The credential
        cannot be used as a user JWT or to read arbitrary AgentRuns.
    UserBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    TelemetrySinkProjectionPublicId:
      name: sink_public_id
      in: path
      required: true
      schema:
        type: string
        pattern: '^tsk_[0-9a-f]{32}$'
    AdapterHandshakeHeader:
      name: DuckDock-Handshake-Id
      in: header
      required: false
      description: Optional G2 binding to an active Generic OTLP adapter handshake.
      schema:
        type: string
        pattern: '^hs_[0-9a-f]{32}$'
...
    NamespaceId:
      name: namespace_id
      in: query
      required: true
      schema:
        type: integer
        minimum: 1
...
  responses:
    Unauthorized:
      description: Missing, invalid, expired, or incorrectly scoped bearer token.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: The authenticated credential lacks the required execution scope.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource does not exist or is not visible in the caller scope.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Invalid lifecycle transition or Idempotency-Key payload conflict.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    ValidationError:
      description: Request violates the v2 lifecycle, correlation, or metadata-only contract.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    OtlpTraceExportRequest:
      type: object
      description: |
        Standard OTLP JSON ExportTraceServiceRequest after Collector
        authentication, governance replacement and metadata-only filtering.
        DuckDock validates correlation metadata but never stores this document.
      required:
        - resourceSpans
      properties:
        resourceSpans:
          type: array
          items:
            type: object
      additionalProperties: true
    OtlpTraceExportResponse:
      type: object
      additionalProperties: false
      properties:
        partialSuccess:
          type: object
          additionalProperties: false
          required:
            - rejectedSpans
            - errorMessage
          properties:
            rejectedSpans:
              type: string
              pattern: '^[0-9]+$'
            errorMessage:
              type: string
              maxLength: 100
...
    AdapterProfile:
      type: string
      enum:
        - openclaw-reporter
        - hermes-reporter
        - generic-otlp-bridge
        - pack-atif-import
    AdapterConfigDrift:
      type: string
      enum:
        - NONE
        - CONFIG_CHANGED
        - BOOT_CHANGED
        - CAPABILITY_CHANGED
    AdapterDescriptor:
      type: object
      additionalProperties: false
      required:
        - adapter_id
        - adapter_version
        - profile
        - source_schema
        - source_schema_version
        - capabilities
        - content_capture_modes
        - instance_id
        - boot_id
        - client_nonce
        - client_time
      properties:
        adapter_id:
          type: string
          minLength: 1
          maxLength: 100
        adapter_version:
          type: string
          minLength: 1
          maxLength: 50
        profile:
          $ref: '#/components/schemas/AdapterProfile'
        protocol:
          type: string
          const: duckdock-adapter
          default: duckdock-adapter
        protocol_version:
          type: string
          const: '1.0'
          default: '1.0'
        source_schema:
          type: string
          minLength: 1
          maxLength: 100
        source_schema_version:
          type: string
          minLength: 1
          maxLength: 50
        capabilities:
          type: array
          minItems: 1
          maxItems: 16
          uniqueItems: true
          items:
            type: string
            pattern: '^[a-z][a-z0-9._-]*$'
        content_capture_modes:
          type: array
          minItems: 1
          maxItems: 1
          items:
            $ref: '#/components/schemas/MetadataOnlyCaptureMode'
        instance_id:
          type: string
          minLength: 1
          maxLength: 128
        boot_id:
          type: string
          minLength: 1
          maxLength: 128
        client_nonce:
          type: string
          minLength: 22
          maxLength: 171
          pattern: '^[A-Za-z0-9_-]+$'
        client_time:
          type: string
          format: date-time
        claimed_capability_level:
          type: string
          enum:
            - DD-C0
            - DD-C1
            - DD-C2
            - DD-C3
          default: DD-C0
        config_fingerprint:
          type:
            - string
            - 'null'
          pattern: '^[0-9a-f]{64}$'
    AdapterHandshake:
      type: object
      additionalProperties: false
      required:
        - handshake_id
        - protocol_version
        - runtime_public_id
        - profile
        - accepted_capabilities
        - rejected_capabilities
        - claimed_capability_level
        - certified_capability_level
        - content_capture_mode
        - clock_skew_degraded
        - limits
        - idempotency
        - server_time
        - expires_at
      properties:
        handshake_id:
          type: string
          pattern: '^hs_[0-9a-f]{32}$'
        protocol_version:
          type: string
          const: '1.0'
        runtime_public_id:
          type: string
          pattern: '^rt_[0-9a-f]{32}$'
        profile:
          $ref: '#/components/schemas/AdapterProfile'
        accepted_capabilities:
          type: array
          maxItems: 16
          items:
            type: string
        rejected_capabilities:
          type: array
          maxItems: 16
          items:
            type: string
        claimed_capability_level:
          type: string
        certified_capability_level:
          type: string
        content_capture_mode:
          $ref: '#/components/schemas/MetadataOnlyCaptureMode'
        clock_skew_degraded:
          type: boolean
        limits:
          type: object
          additionalProperties:
            type: integer
        idempotency:
          type: object
          additionalProperties:
            oneOf:
              - type: integer
              - type: string
        server_time:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
    AdapterHeartbeat:
      type: object
      additionalProperties: false
      required:
        - handshake_id
        - boot_id
        - accepted_capabilities
      properties:
        handshake_id:
          type: string
          pattern: '^hs_[0-9a-f]{32}$'
        boot_id:
          type: string
          minLength: 1
          maxLength: 128
        status:
          type: string
          enum: [ok, degraded, error]
          default: ok
        accepted_capabilities:
          type: array
          minItems: 0
          maxItems: 16
          uniqueItems: true
          items:
            type: string
        config_fingerprint:
          type:
            - string
            - 'null'
          pattern: '^[0-9a-f]{64}$'
        collector_status:
          type:
            - string
            - 'null'
          enum:
            - healthy
            - degraded
            - unavailable
            - not_applicable
            - null
        collector_version:
          type:
            - string
            - 'null'
          maxLength: 64
    AdapterHeartbeatAck:
      type: object
      additionalProperties: false
      required:
        - handshake_id
        - status
        - config_drift
        - rehandshake_required
        - last_heartbeat_at
        - expires_at
      properties:
        handshake_id:
          type: string
          pattern: '^hs_[0-9a-f]{32}$'
        status:
          type: string
          enum: [ACTIVE, DEGRADED, EXPIRED]
        config_drift:
          $ref: '#/components/schemas/AdapterConfigDrift'
        rehandshake_required:
          type: boolean
        last_heartbeat_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
    FleetHeartbeat:
      type: object
      additionalProperties: false
      required:
        - status
        - config_drift
        - collector_status
        - collector_version
        - observed_at
      properties:
        status:
          type: string
        config_drift:
          $ref: '#/components/schemas/AdapterConfigDrift'
        collector_status:
          type:
            - string
            - 'null'
        collector_version:
          type:
            - string
            - 'null'
        observed_at:
          type: string
          format: date-time
    FleetRuntime:
      type: object
      additionalProperties: false
      required:
        - runtime_public_id
        - provider
        - name
        - runtime_status
        - accepted_capabilities
        - rejected_capabilities
        - handshake_status
        - heartbeat_state
        - heartbeat_history
        - namespace_telemetry_sink_count
        - namespace_active_telemetry_sink_count
        - pending_pack_import_count
        - quarantined_item_count
      properties:
        runtime_public_id:
          type: string
          pattern: '^rt_[0-9a-f]{32}$'
        provider:
          type: string
        name:
          type: string
        runtime_status:
          type: string
          enum: [active, degraded, disabled]
        profile:
          oneOf:
            - $ref: '#/components/schemas/AdapterProfile'
            - type: 'null'
        adapter_id:
          type: [string, 'null']
        adapter_version:
          type: [string, 'null']
        certified_capability_level:
          type: [string, 'null']
        accepted_capabilities:
          type: array
          maxItems: 16
          items:
            type: string
        rejected_capabilities:
          type: array
          maxItems: 16
          items:
            type: string
        handshake_status:
          type: string
          enum: [ACTIVE, DEGRADED, EXPIRED, SUPERSEDED, NONE]
        heartbeat_state:
          type: string
          enum: [HEALTHY, STALE, NEVER, ERROR]
        last_heartbeat_at:
          type: [string, 'null']
          format: date-time
        handshake_expires_at:
          type: [string, 'null']
          format: date-time
        config_drift:
          oneOf:
            - $ref: '#/components/schemas/AdapterConfigDrift'
            - type: 'null'
        collector_status:
          type: [string, 'null']
        collector_version:
          type: [string, 'null']
        heartbeat_history:
          type: array
          maxItems: 20
          items:
            $ref: '#/components/schemas/FleetHeartbeat'
        namespace_telemetry_sink_count:
          type: integer
          minimum: 0
        namespace_active_telemetry_sink_count:
          type: integer
          minimum: 0
        latest_run_at:
          type: [string, 'null']
          format: date-time
        pending_pack_import_count:
          type: integer
          minimum: 0
        quarantined_item_count:
          type: integer
          minimum: 0
    FleetSummary:
      type: object
      additionalProperties: false
      required:
        - namespace_id
        - runtime_count
        - healthy_count
        - stale_count
        - degraded_count
        - drifted_count
        - runtimes
      properties:
        namespace_id:
          type: integer
          minimum: 1
        runtime_count:
          type: integer
          minimum: 0
        healthy_count:
          type: integer
          minimum: 0
        stale_count:
          type: integer
          minimum: 0
        degraded_count:
          type: integer
          minimum: 0
        drifted_count:
          type: integer
          minimum: 0
        runtimes:
          type: array
          maxItems: 200
          items:
            $ref: '#/components/schemas/FleetRuntime'
...
    MetadataOnlyCaptureMode:
      type: string
      const: metadata_only
      description: |
        The Reporter control envelope carries only allowlisted identifiers,
        lifecycle timestamps, counters, classified errors, and trace references.
...
    Problem:
      type: object
      additionalProperties: false
      required:
        - status
        - code
        - detail
        - request_id
        - retryable
      properties:
        type:
          type: string
          format: uri-reference
          default: about:blank
        title:
          type: string
          maxLength: 200
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          minLength: 1
          maxLength: 100
        detail:
          type: string
          minLength: 1
          maxLength: 4000
        request_id:
          type: string
          minLength: 1
        retryable:
          type: boolean
        details:
          type: object
          maxProperties: 6
          propertyNames:
            type: string
            enum:
              - operation
              - field
              - limit
              - retry_after_seconds
              - dependency
              - conflict_type
          additionalProperties:
            oneOf:
              - type: string
                maxLength: 500
              - type: number
              - type: boolean
        errors:
          type: array
          items:
            type: object
            additionalProperties: false
            required:
              - location
              - message
            properties:
              location:
                type: string
                minLength: 1
              message:
                type: string
                minLength: 1